tg-bot/requirements.txt
tzdata
typing-extensions
propcache
multidict
magic-filter
idna
frozenlist
certifi
attrs
annotated-types
aiohappyeyeballs
aiofiles
yarl
tzlocal
typing-inspection
pydantic-core
aiosignal
pydantic
apscheduler
aiohttp
aiogram
python-decouple
apscheduler
asyncpg
python-decouple